+++ back/src/main/java/com/example/_thecore_back/emulator/controller/dto/EmulatorResponse.java
@@ -7,8 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class EmulatorResponse {
-
-    private Long id;            // 애뮬레이터 id
+    private int id;            // 애뮬레이터 id
     private String carNumber;   // 차량 번호
     private EmulatorStatus status;      // 상태
 }
