+++ back/src/main/java/com/example/_thecore_back/emulator/controller/dto/RegisterEmulatorResponseData.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 @Builder
 @AllArgsConstructor
 public class RegisterEmulatorResponseData {
-    private Long emulatorId;
+    private int emulatorId;
     private String carNumber;
     private EmulatorStatus emulatorStatus;
 }
