+++ back/src/main/java/com/example/_thecore_back/emulator/controller/dto/UpdateEmulatorResponseData.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @Builder
 public class UpdateEmulatorResponseData {
-    private Long emulatorId;
+    private int emulatorId;
     private String carNumber;
 }
